@@ -7,7 +7,7 @@ issue_tracker: https://github.com/xenoken/udp/issues
 author: Ken Ekeoha <ekeohak@gmail.com>
 
 environment:
-  sdk: '>=2.2.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 #dependencies:
 #  path: ^1.4.1
