@@ -52,7 +52,7 @@ void main() async {
 
   // receiving\listening
   await receiver.listen((datagram) {
-    var str = String.fromCharCodes(datagram.data);
+    var str = String.fromCharCodes(datagram!.data);
     stdout.write(str);
   }, timeout: Duration(seconds: 20));
 
